Minimax AI built and added (proj finished for now)
Employed the use of the minimax algorithm to build a computer player that will always tend towards the best move. Significant delay in first move as it paths out all gamestates, but results in a human vs AI experience that can only end in a tie. Also, general bug fixing and cleaning up some imports.